ref(api): ApiError submodule
